+++ src/spanner/index.ts
@@ -7,5 +7,7 @@ export {
 } from './database.js';
 export { SpannerEntityManager } from './entity-manager.js';
 export * from './errors.js';
+export { SpannerHealthIndicator } from './healthcheck.js';
+export { SpannerModule } from './module.js';
 export { SpannerTable } from './table.decorator.js';
 export * from './types.js';