+++ CHANGELOG.md
@@ -15,3 +15,4 @@ Features:
 - Implement the `AppCheckGuard`.
 - Implement the Firestore utilities, including testing and dependency injection for NestJS.
 - Implement the `SpannerEntityManager` and related utilities.
+- Implement the `SpannerModule` and `SpannerHealthIndicator`.